% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disMat.R
\name{distMat}
\alias{distMat}
\title{Distance matrix}
\usage{
distMat(trees, tipOrder)
}
\arguments{
\item{trees}{Phylogeny trees}

\item{tipOrder}{The names of tip label. You cold give an order to them.}
}
\description{
This function is to get the distance vector of a phylogeny tree in tropical space
}
