% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_pi.R
\name{project_pi}
\alias{project_pi}
\title{Porjected points}
\usage{
project_pi(D_s, D)
}
\arguments{
\item{D_s}{The distance matrix used to build tropical space}

\item{D}{Distance vector (matrix)}
}
\description{
Porjected points
}
