% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tropMCMC.R
\name{tropMCMC}
\alias{tropMCMC}
\title{Tropical MCMC}
\usage{
tropMCMC(distVect_all, N, pcs, nr, env = .GlobalEnv)
}
\arguments{
\item{distVect_all}{}

\item{N}{Number of repeat times.}

\item{pcs}{Number of pricinpal components}

\item{nr}{Number of points in tropical space.}

\item{env}{Parameter for changing environment.}
}
\description{
Tropical MCMC
}
